pyproject.toml
[tool.poetry]
name = "apifastapitemplate"
version = "0.1.1"
description = ""
authors = ["Diego Vásquez <dhvasquez@uc.cl>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.26.0"
uvicorn = "^0.16.0"
fastapi = "^0.70.1"
python-dotenv = "^0.19.2"
pydantic = {extras = ["email"], version = "^1.8.2"}
pytz = "^2021.3"
motor = "^2.5.1"
Jinja2 = "^3.0.3"
beanie = "^1.8.9"

[tool.poetry.dev-dependencies]
mypy = "^0.930"
flake8 = "^4.0.1"
bandit = "^1.7.2"
pytest = "^7.0.1"
pytest-asyncio = "^0.18.1"
httpx = "^0.22.0"
trio = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
